menu_keys: Move button handling from menu.py to new file menu_keys.py
